+++ CHANGES.md
@@ -8,6 +8,7 @@ _Released TDB_
 - Add ability to control showing rotted and same-origin results for root entrypoint and simple embed
 - Add readme sections to explain new filtering and embed features
 - Ensure the `date_added` field is expressed in UTC
+- Fix error when linkrot checking with an ID that does not exist in the webring
 - Replace `requests` with `python-httpx`
 - Update OpenAPI spec version
 - Update minimum Python version to 3.11